+++ operators/decayoperator.cpp
@@ -9,8 +9,8 @@ DecayOperator::DecayOperator() :
 DecayOperator::~DecayOperator() {
 }
 
-void DecayOperator::apply(Particle *p, int elapsed) {
-  if(p->lifetime < 0) {
+void DecayOperator::apply(Particle *p, float elapsed) {
+  if(p->lifetime < 0.0) {
     p->dead = true;
   }
 }
